+++ skfda/__init__.py
@@ -27,7 +27,7 @@ from .representation import FData
 from .representation import FDataBasis
 from .representation import FDataGrid
 
-from . import representation, datasets, preprocessing, exploratory, math
+from . import representation, datasets, preprocessing, exploratory, misc
 
 import os as _os
