+++ skfda/misc/_math.py
