+++ skfda/misc/covariances.py
