+++ skfda/misc/kernels.py
