+++ skfda/misc/metrics.py
